+ input based motor controll

+++ src/Battery.h
@@ -1,2 +1,5 @@
+// returns the voltage of a battery as percentage
 int BatteryVoltageInPercent();
-int BatteryVoltage();
+
+// returns the voltage of a battery as float 
+float BatteryVoltage();
